Restructure the package structure
