+++ app/src/main/java/dev/patrickgold/florisboard/common/NativeInstanceWrapper.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.ime.core
+package dev.patrickgold.florisboard.common
 
 /**
  * Type alias for a native pointer.
