+++ app/src/main/java/dev/patrickgold/florisboard/common/Pointer.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.ime.view
+package dev.patrickgold.florisboard.common
 
 import androidx.annotation.RestrictTo
