+++ app/src/main/java/dev/patrickgold/florisboard/ime/popup/PopupExtendedView.kt
@@ -30,7 +30,7 @@ import dev.patrickgold.florisboard.R
 import dev.patrickgold.florisboard.ime.keyboard.Key
 import dev.patrickgold.florisboard.ime.theme.Theme
 import dev.patrickgold.florisboard.ime.theme.ThemeManager
-import dev.patrickgold.florisboard.ime.view.ViewUtils
+import dev.patrickgold.florisboard.common.ViewUtils
 import kotlin.math.min
 
 class PopupExtendedView : View, ThemeManager.OnThemeUpdatedListener {
