+++ app/src/main/res/layout/smartbar.xml
@@ -8,7 +8,7 @@
     android:orientation="horizontal"
     android:layoutDirection="locale">
 
-    <dev.patrickgold.florisboard.ime.core.FlorisViewFlipper
+    <dev.patrickgold.florisboard.common.FlorisViewFlipper
         android:id="@+id/action_start_area"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
@@ -28,9 +28,9 @@
             android:contentDescription="@string/smartbar__quick_action__exit_editing"
             android:src="@drawable/ic_arrow_back"/>
 
-    </dev.patrickgold.florisboard.ime.core.FlorisViewFlipper>
+    </dev.patrickgold.florisboard.common.FlorisViewFlipper>
 
-    <dev.patrickgold.florisboard.ime.core.FlorisViewFlipper
+    <dev.patrickgold.florisboard.common.FlorisViewFlipper
         android:id="@+id/main_area"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
@@ -111,9 +111,9 @@
             android:layoutDirection="ltr"
             app:isSmartbarKeyboard="true"/>
 
-    </dev.patrickgold.florisboard.ime.core.FlorisViewFlipper>
+    </dev.patrickgold.florisboard.common.FlorisViewFlipper>
 
-    <dev.patrickgold.florisboard.ime.core.FlorisViewFlipper
+    <dev.patrickgold.florisboard.common.FlorisViewFlipper
         android:id="@+id/action_end_area"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
@@ -127,6 +127,6 @@
             android:contentDescription="@string/smartbar__quick_action__private_mode"
             android:src="@drawable/ic_security"/>
 
-    </dev.patrickgold.florisboard.ime.core.FlorisViewFlipper>
+    </dev.patrickgold.florisboard.common.FlorisViewFlipper>
 
 </dev.patrickgold.florisboard.ime.text.smartbar.SmartbarView>
